pyproject.toml
[project]
name = "vid-subs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20240930",
]
